Feature/ms 78 document for module management

+++ MiSurvey-backend/src/services/module.service.js
@@ -12,9 +12,15 @@ const createModule = async (moduleData) => {
 
 const updateModule = async (id, moduleData) => {
   try {
-    const [updatedRows] = await Module.update(moduleData, {
-      where: { ModuleID: id },
-    });
+    const module = await Module.findByPk(id);
+    if (!module) {
+      return {
+        status: false,
+        message: "Module not found",
+      };
+    }
+
+    const [updatedRows] = await Module.update(moduleData, { where: { ModuleID: id } });
     if (updatedRows === 0) {
       return { status: false, message: "No module updated" };
     }
@@ -26,8 +32,16 @@ const updateModule = async (id, moduleData) => {
 
 const deleteModule = async (id) => {
   try {
+    const module = await Module.findByPk(id);
+
+    if (!module) {
+      return { status: false, message: "Module not found" };
+    }
+
     await Module.destroy({ where: { ModuleID: id } });
+
     return { status: true, message: "Module deleted successfully" };
+
   } catch (error) {
     return { status: false, message: error.message, error: error.toString() };
   }
@@ -45,10 +59,13 @@ const getAllModules = async () => {
 const getOneModule = async (id) => {
   try {
     const module = await Module.findByPk(id);
+
     if (!module) {
       return { status: false, message: "Module not found" };
     }
+
     return { status: true, message: "Module fetched successfully", module };
+    
   } catch (error) {
     return { status: false, message: error.message, error: error.toString() };
   }
